src/requirements.txt
aioredis
asgiref
async-timeout
attrs
autobahn
Automat
bleach
bokeh
boto3
certifi
cffi
channels
chardet
constantly
cryptography
daphne
Django
django-bootstrap-static
django-cropperjs
django-storages
gunicorn
hiredis
hyperlink
idna
importlib-metadata
incremental
lxml
msgpack
numpy
opencv-python
optional-django
packaging
panel
param
psycopg
psycopg-binary
pyasn1
pyasn1-modules
pycparser
PyHamcrest
pyOpenSSL
pyparsing
python-decouple
pytz
redis
requests
service-identity
six
sqlparse
Twisted
txaio
urllib3
webencodings
zipp
zope.interface
